[Decomp][Player] Cohesion split: move Player taxi-flight methods into PlayerTaxi.cpp
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  PlayerTaxi::InitTaxiNodes,PlayerTaxi::InitTaxiNodesForClass,PlayerTaxi::InitTaxiNodesForRace,PlayerTaxi::InitTaxiNodesForFaction,PlayerTaxi::InitTaxiNodesForLvl,PlayerTaxi::GetFlightMasterFactionTemplate
PlayerTaxi::InitTaxiNodes,PlayerTaxi::InitTaxiNodesForClass,PlayerTaxi::InitTaxiNodesForRace,PlayerTaxi::InitTaxiNodesForFaction,PlayerTaxi::InitTaxiNodesForLvl,PlayerTaxi::GetFlightMasterFactionTemplate
Ported-from: mangosthree/server@b5281f5867cb
